provide basic responder functionality without serde, json features with serde feature
