+++ src/lib.rs
@@ -4,7 +4,6 @@ pub mod extractors;
 #[cfg(feature = "guards")]
 pub mod guard;
 pub mod headers;
-#[cfg(feature = "responders")]
 pub mod responders;
 
 use axum::{
@@ -15,5 +14,4 @@ pub use extractors::*;
 #[cfg(feature = "guards")]
 pub use guard::*;
 pub use headers::*;
-#[cfg(feature = "responders")]
 pub use responders::*;
